+++ NEWS.md
@@ -1,6 +1,7 @@
 # fitur 0.6.1
 
 * added test function for distfun objects
+
 * diagnostic plots now have better checks for distfun objects and lists of
 distfun objects