+++ NEWS.md
@@ -1,5 +1,7 @@
 # fitur 0.6.1
 
+* shiny gadget for fitting univariate distributions has been added
+
 * added test function for distfun objects
 
 * diagnostic plots now have better checks for distfun objects and lists of